Revert "debugging, fix to rounding error in calculating delta time from scale factors"
This reverts commit 388c4ea3bcb832ee6e986bfc90d2f55b966bad07.

+++ src/runner.c
@@ -375,7 +375,7 @@ void runner_do_stars_ghost(struct runner *r, struct cell *c, int timer) {
           double star_age, current_time_begin = -1;
           if (with_cosmology) {
             star_age = cosmology_get_delta_time_from_scale_factors(
-                cosmo, sp->birth_scale_factor, (float) cosmo->a); // use float to avoid rounding errors
+                cosmo, sp->birth_scale_factor, cosmo->a);
           } else {
             current_time_begin =
                 get_integer_time_begin(e->ti_current - 1, sp->time_bin) *
